Use uint32_t for samples per pixel

+++ iris_camera_toolkit/halton_image_sampler.c
@@ -428,7 +428,7 @@ HaltonImageSamplerDuplicate(
 
 ISTATUS
 HaltonImageSamplerAllocate(
-    _In_ uint16_t samples_per_pixel,
+    _In_ uint32_t samples_per_pixel,
     _Out_ PIMAGE_SAMPLER *image_sampler
     )
 {
